changes to test (partly for upcoming commits)

+++ test-compiler/0_compiler/universe/universe-default.m
@@ -4,7 +4,7 @@
 int  main( void)
 {
    printf( "name : %s\n",  __MULLE_OBJC_UNIVERSENAME__ ? __MULLE_OBJC_UNIVERSENAME__ : "NULL");
-   printf( "id   : %lx\n", __MULLE_OBJC_UNIVERSEID__);
+   printf( "id   : %lx\n", (long) __MULLE_OBJC_UNIVERSEID__);
 
    return( 0);
 }
